Added todo comments for the google envs and the personalGoogleAuth.
Made some changes to the service names so they represent the logic better.

Removed the function to create a Pdf because its not needed.

Changed the return data of the get translated memo pdf to return the actual pdf stream when the handler is called (previously it returned the data as an JSON body containing all of the data of the memoRecord)